Add Hatz poster/trash bag interaction
Now, when the final trash bag is deposited, a mysterious pterosaur will knock the bag out of the dumpster, splitting it open, revealing
campaign posters for Hatz Beakman.

In the future this will include a crowbar to
allow you to traverse the sewers and also to
further incriminate the mayor.